Split monolithic proposal script into modular files
