+++ proposal_generator.py
@@ -4,7 +4,6 @@ import holidays
 import PyPDF2
 import tkinter as tk
 from tkinter import ttk, filedialog, messagebox
-from datetime import datetime, timedelta
 import json
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
@@ -30,32 +29,16 @@ import csv
 import io
 import re
 import math
-# --- MODIFICATION: Added import for openpyxl ---
-try:
-    from openpyxl import Workbook, load_workbook
-    from openpyxl.styles import Font
-    from openpyxl.utils import get_column_letter
-except ImportError:
-    messagebox.showerror("Missing Dependency", "The 'openpyxl' library is required to work with Excel files. Please install it using: pip install openpyxl")
-    sys.exit()
+from io import BytesIO
+
 try:
     from tkcalendar import DateEntry
     HAS_TKCAL = True
 except ImportError:
     HAS_TKCAL = False
 
-# --- HELPER FUNCTION FOR PYINSTALLER ---
-# This function helps find bundled files (like fonts and logos) when running as an .exe
-def resource_path(relative_path):
-    """ Get absolute path to resource, works for dev and for PyInstaller """
-    try:
-        # PyInstaller creates a temp folder and stores path in _MEIPASS
-        base_path = sys._MEIPASS
-    except Exception:
-        base_path = os.path.abspath(".")
-
-    return os.path.join(base_path, relative_path)
-
+from utils import resource_path
+from gantt import build_gantt_with_version
 class ProposalItem:
     """Represents a single task or milestone in the project."""
     def __init__(self, name, duration=0, price=0, start_date="", is_milestone=False, indent_level=0, item_id=None):
@@ -76,317 +59,6 @@ class ProposalItem:
         self.lag = 0 # Lag in days
         # --- MODIFICATION: Add flag for manually set start dates ---
         self.is_start_pinned = False
-
-
-# Gantt Chart Constants
-FONTSIZE_TABLE = 8
-FONTSIZE_CHART_REGULAR = 7.5
-FONTSIZE_CHART_SUMMARY = 7.5
-FONTSIZE_XTICK = 7.0
-MAX_ROWS_PER_PAGE = 35
-PRIMARY = "#991f2b"
-SECONDARY = "black"
-LEFT_RIGHT_WIDTHS = [1.6, 2.4]
-COL_EDGES = [0.00, 0.65, 0.75, 0.87, 1.00]
-HEADERS = ["Task", "Duration", "Start", "Finish"]
-MAX_NAME_LENGTH = 50
-ROW_HEIGHT = 1.0
-MIN_WIDTH_DAYS = 0.2
-
-# Replace all the current Gantt functions with these from testgantt.py:
-
-def _nsmap(root):
-    """Detect default namespace from the root tag and return a dict for XPath."""
-    m = re.match(r"\{(.*)\}", root.tag)
-    uri = m.group(1) if m else ""
-    return {"msp": uri} if uri else {}
-
-def _txt(el, tag, ns):
-    """Extract text from element, handling both namespaced and non-namespaced tags."""
-    # Try with namespace first
-    if ns and "msp" in ns:
-        t = el.find(f"msp:{tag}", ns)
-        if t is not None and t.text:
-            return t.text.strip()
-    
-    # Try without namespace
-    t = el.find(tag)
-    if t is not None and t.text:
-        return t.text.strip()
-    
-    return None
-
-def _parse_dt(s):
-    """Parse ISO-like datetime without external deps."""
-    if not s:
-        return None
-    # Handles strings like "2025-08-01T08:00:00" (ignore any trailing Z)
-    try:
-        return datetime.fromisoformat(s.replace("Z", ""))
-    except ValueError:
-        # If that fails, try other common formats
-        try:
-            return datetime.strptime(s, "%Y-%m-%dT%H:%M:%S")
-        except ValueError:
-            return None
-
-def compute_duration_days(start, finish):
-    """Integer calendar-day span; same-day shows as 0."""
-    if not start or not finish:
-        return None
-    return max(0, (finish.date() - start.date()).days)
-
-def estimate_text_width_in_days(text, fontsize, chart_span_days, chart_width_inches):
-    """
-    Estimate the width of text in chart coordinate system (days).
-    Balanced estimation for proper gaps without overlap.
-    """
-    # Balanced estimate: each character is about 0.6 * fontsize points wide
-    char_width_points = fontsize * 0.6
-    char_width_inches = char_width_points / 72
-    text_width_inches = len(text) * char_width_inches
-    
-    # Convert inches to days based on chart proportions
-    days_per_inch = chart_span_days / chart_width_inches
-    text_width_days = text_width_inches * days_per_inch
-    
-    # Keep 10% safety margin for readability
-    return text_width_days * 1.1
-
-def _build_one_page_with_version(pdf, page_rows, page_num, total_pages, x_min, x_max, true_start, 
-                    total_span_days, chart_width_inches, title, project_title, customer_name, logo_path, version="V1"):
-    """Renders a single page of the Gantt chart with version info."""
-    header_row = {
-        "name": "Task", "start": None, "finish": None, "kind": "header", "dur": "Duration"
-    }
-    all_rows_for_page = [header_row] + page_rows
-
-    fig = plt.figure(figsize=(16, 10))
-
-    # Manual Axes Creation for Fixed Row Height
-    left_margin, right_margin, bottom_margin, top_margin = 0.03, 0.99, 0.10, 0.88
-    plot_area_w = right_margin - left_margin
-    full_plot_area_h = top_margin - bottom_margin
-    rows_on_full_page = MAX_ROWS_PER_PAGE + 1
-    rows_on_this_page = len(all_rows_for_page)
-    height_ratio = rows_on_this_page / rows_on_full_page
-    current_plot_h = full_plot_area_h * height_ratio
-    current_bottom = top_margin - current_plot_h
-    table_w = plot_area_w * (LEFT_RIGHT_WIDTHS[0] / sum(LEFT_RIGHT_WIDTHS))
-    chart_w = plot_area_w * (LEFT_RIGHT_WIDTHS[1] / sum(LEFT_RIGHT_WIDTHS))
-    table_l, chart_l = left_margin, left_margin + table_w
-    ax_left = fig.add_axes([table_l, current_bottom, table_w, current_plot_h])
-    ax_right = fig.add_axes([chart_l, current_bottom, chart_w, current_plot_h])
-    ax_right.sharey(ax_left)
-
-    # LEFT: table
-    ax_left.set_xlim(0, 1)
-    ax_left.set_ylim(-0.5, len(all_rows_for_page) - 0.5)
-    ax_left.invert_yaxis()
-    ax_left.axis("off")
-
-    for idx, r in enumerate(all_rows_for_page):
-        is_header_row = r["kind"] == "header"
-        is_summary_row = r["kind"] == "summary"
-        for c in range(4):
-            x0 = COL_EDGES[c]
-            w = COL_EDGES[c + 1] - COL_EDGES[c]
-            if is_header_row or is_summary_row:
-                ax_left.add_patch(Rectangle((x0, idx - 0.5), w, 1.0, facecolor=PRIMARY, 
-                                            edgecolor=SECONDARY, linewidth=1.2 if is_header_row else 0.6))
-            else:
-                ax_left.add_patch(Rectangle((x0, idx - 0.5), w, 1.0, fill=False, 
-                                            edgecolor=SECONDARY, linewidth=0.6))
-        if is_header_row:
-            for c, header_text in enumerate(HEADERS):
-                center_x = (COL_EDGES[c] + COL_EDGES[c + 1]) / 2
-                ax_left.text(center_x, idx, header_text, va="center", ha="center", 
-                             fontsize=FONTSIZE_TABLE, fontweight="bold", color="white")
-        else:
-            name_text = r["name"].replace('\n', ' ').replace('\r', ' ')
-            if len(name_text) > MAX_NAME_LENGTH:
-                name_text = name_text[:MAX_NAME_LENGTH-3] + "..."
-            text_color = "white" if is_summary_row else SECONDARY
-            font_weight = "bold" if is_summary_row else "normal"
-            ax_left.text(COL_EDGES[0] + 0.008, idx, name_text, va="center", ha="left", 
-                         fontsize=FONTSIZE_TABLE, fontweight=font_weight, color=text_color)
-            ax_left.text((COL_EDGES[1] + COL_EDGES[2]) / 2, idx, 
-                         ("" if r["dur"] is None else f"{r['dur']}d"), 
-                         va="center", ha="center", fontsize=FONTSIZE_TABLE, color=text_color)
-            if r["start"]:
-                ax_left.text((COL_EDGES[2] + COL_EDGES[3]) / 2, idx, 
-                             r["start"].strftime("%m/%d/%y"), va="center", ha="center", fontsize=FONTSIZE_TABLE, color=text_color)
-            if r["finish"]:
-                ax_left.text((COL_EDGES[3] + COL_EDGES[4]) / 2, idx, 
-                             r["finish"].strftime("%m/%d/%y"), va="center", ha="center", fontsize=FONTSIZE_TABLE, color=text_color)
-    
-    ax_left.axhline(y=-0.5, linestyle="-", linewidth=2.0, color=SECONDARY)
-    ax_left.axhline(y=len(all_rows_for_page) - 0.5, linestyle="-", linewidth=2.0, color=SECONDARY)
-    ax_left.axvline(x=0, linestyle="-", linewidth=2.0, color=SECONDARY)
-
-    # RIGHT: chart
-    chart_right_edge = mdates.date2num(x_max)
-    ax_right.set_xlim(mdates.date2num(x_min), chart_right_edge)
-    ax_right.xaxis.tick_top()
-    ax_right.xaxis.set_label_position('top')
-    project_duration_months = (x_max.year - x_min.year) * 12 + (x_max.month - x_min.month)
-    ax_right.xaxis.set_major_locator(MonthLocator(interval=2 if project_duration_months > 24 else 1))
-    ax_right.xaxis.set_major_formatter(DateFormatter("%b %Y"))
-    ax_right.xaxis.set_minor_locator(mticker.NullLocator())
-    ax_right.tick_params(axis='x', which='major', labelsize=FONTSIZE_XTICK, pad=2)
-    for label in ax_right.get_xticklabels():
-        label.set_bbox(dict(boxstyle="round,pad=0.3", facecolor="white", edgecolor=SECONDARY, linewidth=0.8))
-    ax_right.grid(which="major", axis="x", linestyle="--", linewidth=0.8, color=SECONDARY, alpha=0.6)
-    for y in range(len(all_rows_for_page) + 1):
-        ax_right.axhline(y=y - 0.5, linestyle="--", linewidth=0.6, color=SECONDARY, alpha=0.5)
-    ax_right.axhline(y=-0.5, linestyle="-", linewidth=2.0, color=SECONDARY)
-    ax_right.axhline(y=len(all_rows_for_page) - 0.5, linestyle="-", linewidth=2.0, color=SECONDARY)
-    ax_right.axvline(x=chart_right_edge, linestyle="-", linewidth=2.0, color=SECONDARY)
-    ax_right.set_yticks([])
-    ax_right.tick_params(axis='y', which='both', length=0)
-    chart_left_edge = mdates.date2num(x_min)
-    
-    for idx, r in enumerate(all_rows_for_page):
-        if idx == 0: continue
-            
-        if r["kind"] != "summary" and r["start"] and r["finish"]:
-            start_num, finish_num = mdates.date2num(r["start"]), mdates.date2num(r["finish"])
-            is_milestone = (finish_num - start_num) < 0.01
-            
-            if is_milestone:
-                diamond_height = 0.3
-                diamond_width_days = max(3, total_span_days * 0.005) 
-                diamond_x, diamond_y = start_num, idx
-                diamond_verts = [(diamond_x, diamond_y + diamond_height), (diamond_x + diamond_width_days, diamond_y), (diamond_x, diamond_y - diamond_height), (diamond_x - diamond_width_days, diamond_y), (diamond_x, diamond_y + diamond_height)]
-                diamond = plt.Polygon(diamond_verts, facecolor=PRIMARY, edgecolor=SECONDARY, linewidth=1.5)
-                ax_right.add_patch(diamond)
-                name_text = r["name"].replace('\n', ' ').replace('\r', ' ')
-                text_width_days = estimate_text_width_in_days(name_text, FONTSIZE_CHART_SUMMARY, total_span_days, chart_width_inches)
-                text_start_x = diamond_x + diamond_width_days + (total_span_days * 0.005)
-                if chart_right_edge - text_start_x > text_width_days:
-                    ax_right.text(text_start_x, diamond_y, name_text, va="center", ha="left", fontsize=FONTSIZE_CHART_SUMMARY, color=SECONDARY)
-            else:
-                span = max(finish_num - start_num, MIN_WIDTH_DAYS)
-                bar_height, bar_y = 0.5, idx - 0.25
-                ax_right.broken_barh([(start_num, span)], (bar_y, bar_height), facecolors=PRIMARY, edgecolors=SECONDARY, linewidth=0.8)
-                name_text = r["name"].replace('\n', ' ').replace('\r', ' ')
-                text_width_days = estimate_text_width_in_days(name_text, FONTSIZE_CHART_REGULAR, total_span_days, chart_width_inches)
-                margin_days = max(1.5, total_span_days * 0.015)
-                space_on_right = chart_right_edge - finish_num - margin_days
-                space_on_left = start_num - chart_left_edge - margin_days
-                if space_on_right >= text_width_days:
-                    ax_right.text(finish_num + margin_days, idx, name_text, va="center", ha="left", fontsize=FONTSIZE_CHART_REGULAR, color=SECONDARY)
-                elif space_on_left >= text_width_days:
-                    ax_right.text(start_num - margin_days, idx, name_text, va="center", ha="right", fontsize=FONTSIZE_CHART_REGULAR, color=SECONDARY)
-        elif r["kind"] == "summary" and r["start"] and r["finish"]:
-            start_num, finish_num = mdates.date2num(r["start"]), mdates.date2num(r["finish"])
-            span = max(finish_num - start_num, MIN_WIDTH_DAYS)
-            bar_height, bar_y, cap_height = 0.1, idx - 0.05, 0.4
-            name_text = r["name"].replace('\n', ' ').replace('\r', ' ')
-            text_width_days = estimate_text_width_in_days(name_text, FONTSIZE_CHART_SUMMARY, total_span_days, chart_width_inches)
-            margin_days = max(1.5, total_span_days * 0.015)
-            space_on_right = chart_right_edge - (start_num + span) - margin_days
-            space_on_left = start_num - chart_left_edge - margin_days
-            if space_on_right >= text_width_days:
-                ax_right.broken_barh([(start_num, span)], (bar_y, bar_height), facecolors=SECONDARY, edgecolors=SECONDARY, linewidth=1.0)
-                ax_right.plot([start_num, start_num], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                ax_right.plot([start_num + span, start_num + span], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                ax_right.text(start_num + span + margin_days, idx, name_text, va="center", ha="left", fontsize=FONTSIZE_CHART_SUMMARY, color=SECONDARY)
-            elif space_on_left >= text_width_days:
-                ax_right.broken_barh([(start_num, span)], (bar_y, bar_height), facecolors=SECONDARY, edgecolors=SECONDARY, linewidth=1.0)
-                ax_right.plot([start_num, start_num], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                ax_right.plot([start_num + span, start_num + span], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                ax_right.text(start_num - margin_days, idx, name_text, va="center", ha="right", fontsize=FONTSIZE_CHART_SUMMARY, color=SECONDARY)
-            else:
-                bar_center = start_num + span / 2
-                text_space_needed = min(text_width_days + 0.2, span * 0.9)
-                left_split_end = bar_center - text_space_needed / 2
-                right_split_start = bar_center + text_space_needed / 2
-                min_segment_width = max(0.5, total_span_days * 0.005)
-                if left_split_end > start_num and (left_split_end - start_num) >= min_segment_width:
-                    ax_right.broken_barh([(start_num, left_split_end - start_num)], (bar_y, bar_height), facecolors=SECONDARY, edgecolors=SECONDARY, linewidth=1.0)
-                    ax_right.plot([start_num, start_num], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                if right_split_start < start_num + span and ((start_num + span) - right_split_start) >= min_segment_width:
-                    ax_right.broken_barh([(right_split_start, (start_num + span) - right_split_start)], (bar_y, bar_height), facecolors=SECONDARY, edgecolors=SECONDARY, linewidth=1.0)
-                    ax_right.plot([start_num + span, start_num + span], [idx - cap_height/2, idx + cap_height/2], color=SECONDARY, linewidth=2)
-                ax_right.text(bar_center, idx, name_text, va="center", ha="center", fontsize=FONTSIZE_CHART_SUMMARY, color=SECONDARY)
-
-    ax_right.axvline(mdates.date2num(true_start), linestyle="--", linewidth=1.0, color=SECONDARY, alpha=0.4)
-    fig.text(0.5, 0.96, title, ha='center', va='bottom', fontsize=16, fontweight='bold', color=PRIMARY)
-    info_y_start = 0.98
-    if project_title:
-        fig.text(0.03, info_y_start, f"Project: {project_title}", ha='left', va='top', fontsize=11, fontweight='bold', color=PRIMARY)
-    if customer_name:
-        customer_y = info_y_start - 0.025 if project_title else info_y_start
-        fig.text(0.03, customer_y, f"Customer: {customer_name}", ha='left', va='top', fontsize=11, color=PRIMARY)
-    if logo_path:
-        try:
-            logo_img = plt.imread(logo_path)
-            logo_height_fig, right_edge, top_edge = 0.06, 0.99, 0.98
-            aspect_ratio = logo_img.shape[1] / logo_img.shape[0]
-            fig_width_in, fig_height_in = fig.get_size_inches()
-            logo_height_in = logo_height_fig * fig_height_in
-            logo_width_in = logo_height_in * aspect_ratio
-            logo_width_fig = logo_width_in / fig_width_in
-            left_pos = right_edge - logo_width_fig
-            bottom_pos = top_edge - logo_height_fig
-            ax_logo = fig.add_axes([left_pos, bottom_pos, logo_width_fig, logo_height_fig], anchor='NE', zorder=10)
-            ax_logo.imshow(logo_img)
-            ax_logo.axis('off')
-        except Exception as e:
-            print(f"Warning: Could not load or place logo. Error: {e}")
-    if total_pages > 1:
-        fig.text(0.99, 0.01, f'Page {page_num} of {total_pages}', ha='right', va='bottom', fontsize=7, color='gray')
-
-    # Footer with date and version (CHANGED FROM ORIGINAL)
-    date_str = datetime.now().strftime("%B %d, %Y")
-    fig.text(0.03, 0.02, f"{date_str} - {version}", ha='left', va='bottom', fontsize=8, color='gray')
-
-    pdf.savefig(fig, bbox_inches="tight")
-    plt.close(fig)
-
-
-def build_gantt_with_version(rows, out_pdf, title="Project Schedule", project_title="", customer_name="", logo_path="", version="V1"):
-    """Render the Gantt to PDF with version info in footer."""
-    if not rows:
-        print("Warning: No tasks to plot.")
-        return
-        
-    starts = [r["start"] for r in rows if r["start"]]
-    finishes = [r["finish"] for r in rows if r["finish"]]
-    if not starts or not finishes:
-        raise ValueError("No dated tasks found.")
-    true_start, true_finish = min(starts), max(finishes)
-    
-    span_days = max(1, (true_finish.date() - true_start.date()).days)
-    diamond_width_days = max(3, span_days * 0.005)
-    left_pad_days = diamond_width_days + 2
-    right_pad_days = max(1, int(span_days * 0.05))
-    x_min = true_start - timedelta(days=left_pad_days)
-    x_max = true_finish + timedelta(days=right_pad_days)
-    total_span_days = (x_max.date() - x_min.date()).days
-    
-    for r in rows:
-        r["dur"] = compute_duration_days(r["start"], r["finish"])
-
-    with PdfPages(out_pdf) as pdf:
-        total_pages = (len(rows) + MAX_ROWS_PER_PAGE - 1) // MAX_ROWS_PER_PAGE
-        chart_width_inches = 16 * LEFT_RIGHT_WIDTHS[1] / sum(LEFT_RIGHT_WIDTHS)
-
-        for i in range(0, len(rows), MAX_ROWS_PER_PAGE):
-            page_rows_chunk = rows[i : i + MAX_ROWS_PER_PAGE]
-            page_num = (i // MAX_ROWS_PER_PAGE) + 1
-            print(f"Generating page {page_num} of {total_pages}...")
-            _build_one_page_with_version(
-                pdf=pdf, page_rows=page_rows_chunk, page_num=page_num, total_pages=total_pages,
-                x_min=x_min, x_max=x_max, true_start=true_start, total_span_days=total_span_days,
-                chart_width_inches=chart_width_inches, title=title, project_title=project_title,
-                customer_name=customer_name, logo_path=logo_path, version=version
-            )
-from reportlab.platypus import Flowable
-from io import BytesIO
-
-
 class ProposalGenerator:
     """
     The main application class for the PDF Proposal Generator.
@@ -2540,776 +2212,3 @@ class ProposalGenerator:
         except Exception as e:
             messagebox.showerror("Error", f"Failed to load Excel template: {str(e)}")
 
-import os
-import math
-import tkinter as tk
-from tkinter import ttk, filedialog, messagebox
-from datetime import datetime
-import pandas as pd
-
-# Your existing UI/engine
-# from proposal import ProposalGenerator, ProposalItem  # merged above
-# ===================
-# Parsing & Build Rules
-# ===================
-
-EXACT_SUBTOTAL_LABELS = {
-    "civil engineering", "electrical engineering", "structural engineering", "substation engineering"
-}
-
-PHASES = ["30%", "60%", "90%", "IFC"]
-
-
-def _pairs(ncols: int):
-    """(task, price) column pairs on Proposal Page. Includes D/E explicitly."""
-    return [(0, 1), (3, 4), (6, 7), (9, 10), (10, 11)]
-
-
-def _categorize(text: str) -> str:
-    t = (text or "").lower().strip()
-    if t.startswith("civil"): return "Civil"
-    if t.startswith("electrical"): return "Electrical"
-    if t.startswith("structural"): return "Structural"
-    if t.startswith("substation"): return "Substation"
-    # NEW: BESS detection (handles "BESS", "BESS Engineering", "Battery Energy Storage")
-    if t.startswith("bess") or "battery energy storage" in t:
-        return "BESS"
-    return ""
-
-
-
-def _infer_phase(text: str):
-    t = (text or "").lower()
-    if "30% design" in t or "30%" in t: return "30%"
-    if "60% design" in t or "60%" in t: return "60%"
-    if "90% design" in t or "90%" in t: return "90%"
-    if "ifc design" in t or "record drawings" in t or "ifc" in t: return "IFC"
-    return None
-
-
-def load_proposal_page_rows(path: str):
-    pp = pd.read_excel(path, sheet_name="Proposal Page", header=None, engine="openpyxl")
-    props = []
-    ncols = pp.shape[1]
-
-    def normalize(s):
-        return (s or "").strip()
-
-    for (txt_col, price_col) in _pairs(ncols):
-        current_phase = None
-        current_category = None  # tracks the most recent category header
-
-        for i in range(pp.shape[0]):
-            txt = pp.iat[i, txt_col] if txt_col < ncols else None
-            val = pp.iat[i, price_col] if price_col < ncols else None
-
-            # Phase detection
-            if isinstance(txt, str):
-                maybe = _infer_phase(txt)
-                if maybe:
-                    current_phase = maybe
-
-                # Category header detection (now includes Substation & BESS)
-                low = txt.lower().strip()
-                if low in (
-                    "civil engineering",
-                    "electrical engineering",
-                    "structural engineering",
-                    "substation engineering",
-                    "bess",
-                    "bess engineering",
-                    "battery energy storage",
-                    "battery energy storage system",
-                ):
-                    if "bess" in low or "battery energy storage" in low:
-                        current_category = "BESS"
-                    elif low.startswith("substation"):
-                        current_category = "Substation"
-                    else:
-                        # "Civil Engineering" -> "Civil", etc.
-                        current_category = txt.split()[0].capitalize()
-                    continue  # header row itself isn't a task
-
-            # Candidate task row with a numeric price
-            if isinstance(txt, str) and normalize(txt) and pd.notna(val):
-                lower = txt.lower().strip()
-                # Skip totals/headers/etc.
-                if any(k in lower for k in [
-                    "total", "milestone", "summary of services", "engineering proposal", "insurance adder"
-                ]):
-                    continue
-                if lower in EXACT_SUBTOTAL_LABELS:
-                    continue
-
-                # price
-                try:
-                    price = float(val)
-                except Exception:
-                    continue
-
-                # Category assignment: explicit on the row, else the last seen header
-                cat = _categorize(txt) or current_category or ""
-                if not cat:
-                    continue
-
-                props.append({
-                    "category": cat,
-                    "task": normalize(txt),
-                    "proposal_price": price,
-                    "phase": current_phase or _infer_phase(txt) or "30%",
-                })
-    return props
-
-
-
-
-def _load_detail_map(path: str, sheet: str):
-    """Return {Description: {hours, price}}; dedupe by keeping first priced/max price."""
-    try:
-        df = pd.read_excel(path, sheet_name=sheet, header=None, engine="openpyxl")
-    except Exception:
-        return {}
-    df2 = df.iloc[12:].reset_index(drop=True)  # after header row (index 11)
-    DESC, HRS, COST = 2, 11, 12
-    out = {}
-    for _, row in df2.iterrows():
-        desc = row.iloc[DESC] if len(row) > DESC else None
-        hours = row.iloc[HRS] if len(row) > HRS else None
-        price = row.iloc[COST] if len(row) > COST else None
-        if isinstance(desc, str) and desc.strip():
-            name = desc.strip()
-            h = float(hours) if pd.notna(hours) else 0.0
-            p = 0.0
-            if pd.notna(price) and str(price).strip().lower() != "not included":
-                try:
-                    p = float(price)
-                except Exception:
-                    p = 0.0
-            if name not in out:
-                out[name] = {"hours": h, "price": p}
-            else:
-                prev_p = float(out[name].get("price") or 0.0)
-                if (prev_p <= 0 and p > 0) or (p > prev_p):
-                    out[name] = {"hours": h, "price": p}
-    return out
-
-def _load_structural_from_electrical(path: str, sheet: str = "Electrical"):
-    """
-    Parse the 'Structural Engineering' section that lives inside the Electrical sheet.
-    Returns: {task_name: {"hours": float, "price": float}}
-    """
-    try:
-        df = pd.read_excel(path, sheet_name=sheet, header=None, engine="openpyxl")
-    except Exception:
-        return {}
-
-    out = {}
-    DESC, HRS, COST = 2, 11, 12
-
-    # Find the section header (handles misspelling 'Structrural Engineering')
-    header_row = None
-    for i in range(len(df)):
-        row = df.iloc[i]
-        if any(isinstance(v, str) and re.search(r"\bstruct\w*\s+engineering\b", v, re.I) for v in row.tolist()):
-            header_row = i
-            break
-    if header_row is None:
-        return {}
-
-    # Collect rows until a new major section / stage total
-    for i in range(header_row + 1, len(df)):
-        row = df.iloc[i]
-        desc = row.iloc[DESC] if len(row) > DESC else None
-        if isinstance(desc, str) and desc.strip():
-            low = desc.lower().strip()
-            if "stage total" in low or any(k in low for k in ["substation", "bess", "additional services"]):
-                break
-            # skip echoed header lines like "Structural Engineering"
-            if re.search(r"\bstruct\w*\s+engineering\b", low):
-                continue
-
-            hours = row.iloc[HRS] if len(row) > HRS else None
-            price = row.iloc[COST] if len(row) > COST else None
-            h = float(hours) if pd.notna(hours) else 0.0
-            p = 0.0
-            if pd.notna(price) and str(price).strip().lower() != "not included":
-                try:
-                    p = float(price)
-                except Exception:
-                    p = 0.0
-            out[desc.strip()] = {"hours": h, "price": p}
-
-    # Ensure we also capture a standalone "Structural Plan Set" if it appears outside the block
-    for _, row in df.iterrows():
-        desc = row.iloc[DESC] if len(row) > DESC else None
-        if isinstance(desc, str) and "structural plan set" in desc.lower():
-            hours = row.iloc[HRS] if len(row) > HRS else None
-            price = row.iloc[COST] if len(row) > COST else None
-            h = float(hours) if pd.notna(hours) else 0.0
-            p = 0.0
-            if pd.notna(price) and str(price).strip().lower() != "not included":
-                try:
-                    p = float(price)
-                except Exception:
-                    p = 0.0
-            out[desc.strip()] = {"hours": h, "price": p}
-
-    return out
-def _load_design_phase_rows(path: str, prefix: str, sheet: str = "Electrical"):
-    """
-    Pull phase-level rows like 'Substation 60% - Design', 'Substation IFC - Design',
-    or 'BESS 60% - Design' from the Electrical sheet.
-    Returns: {task_name: {"hours": float, "price": float}}
-    """
-    try:
-        df = pd.read_excel(path, sheet_name=sheet, header=None, engine="openpyxl")
-    except Exception:
-        return {}
-
-    out = {}
-    DESC, HRS, COST = 2, 11, 12
-    pref = prefix.lower() + " "
-
-    for _, row in df.iterrows():
-        desc = row.iloc[DESC] if len(row) > DESC else None
-        if isinstance(desc, str) and desc.strip():
-            name = desc.strip()
-            low = name.lower()
-            # e.g., "Substation 60% - Design", "Substation IFC - Design", "BESS 60% - Design"
-            if low.startswith(pref) and "- design" in low:
-                hours = row.iloc[HRS] if len(row) > HRS else None
-                price = row.iloc[COST] if len(row) > COST else None
-                h = float(hours) if pd.notna(hours) else 0.0
-                p = 0.0
-                if pd.notna(price) and str(price).strip().lower() != "not included":
-                    try:
-                        p = float(price)
-                    except Exception:
-                        p = 0.0
-                out[name] = {"hours": h, "price": p}
-    return out
-def enrich_with_details(path: str, rows):
-    """
-    Attach 'hours' and 'detail_price' to each row by looking up the detail sheets.
-    - Electrical tasks: from Electrical sheet
-    - Civil tasks:      from Civil sheet
-    - Structural tasks: from Structural section inside Electrical sheet
-    - Substation tasks: primarily from Civil sheet (e.g., 'Substation Pad Design - Civ. ...')
-    - BESS tasks:       primarily from Electrical sheet (e.g., 'BESS 60% - Design')
-    """
-    xl = pd.ExcelFile(path, engine="openpyxl")
-    civil_map = _load_detail_map(path, "Civil") if "Civil" in xl.sheet_names else {}
-    elec_map  = _load_detail_map(path, "Electrical") if "Electrical" in xl.sheet_names else {}
-    structural_from_elec = _load_structural_from_electrical(path)
-
-    # Helper: tolerant key lookup (handles stray spaces, minor punctuation)
-    import re
-    def _norm(s: str) -> str:
-        return re.sub(r"\s+", " ", (s or "").strip().lower())
-
-    # Precompute normalized maps for fallback matching
-    civil_norm = { _norm(k): v for k, v in civil_map.items() }
-    elec_norm  = { _norm(k): v for k, v in elec_map.items() }
-    struc_norm = { _norm(k): v for k, v in structural_from_elec.items() }
-
-    for r in rows:
-        cat  = (r.get("category") or "").strip()
-        task = (r.get("task") or "").strip()
-        key  = _norm(task)
-
-        # Primary source by category
-        if cat == "Electrical":
-            d = elec_map.get(task) or elec_map.get(task.strip()) or elec_norm.get(key, {})
-        elif cat == "Civil":
-            d = civil_map.get(task) or civil_map.get(task.strip()) or civil_norm.get(key, {})
-        elif cat == "Structural":
-            d = (structural_from_elec.get(task) or structural_from_elec.get(task.strip())
-                 or struc_norm.get(key, {}))
-        elif cat == "Substation":
-            # Substation Pad Design rows live on the Civil sheet
-            d = (civil_map.get(task) or civil_map.get(task.strip()) or civil_norm.get(key, {})
-                 or elec_map.get(task) or elec_map.get(task.strip()) or elec_norm.get(key, {}))
-        elif cat == "BESS":
-            # BESS 60% - Design lives on the Electrical sheet
-            d = (elec_map.get(task) or elec_map.get(task.strip()) or elec_norm.get(key, {})
-                 or civil_map.get(task) or civil_map.get(task.strip()) or civil_norm.get(key, {}))
-        else:
-            d = {}
-
-        r["hours"] = float(d.get("hours")) if d.get("hours") is not None else None
-        r["detail_price"] = float(d.get("price")) if d.get("price") is not None else None
-
-    return rows
-
-
-def extract_project_info(path: str):
-    """
-    Robustly scan the 'Proposal Page' for:
-      - date
-      - client
-      - project
-      - location
-      - state
-      - size_mw  (numeric if possible; otherwise raw string)
-    """
-    df = pd.read_excel(path, sheet_name="Proposal Page", header=None, engine="openpyxl")
-    info = {"date": None, "client": None, "project": None, "location": None, "state": None, "size_mw": None}
-
-    # Keep the old fixed-cell fallbacks if they still apply
-    try:
-        info["date"] = df.iat[0, 1]
-    except Exception:
-        pass
-
-    # Scan top-left area for label:value pairs (avoid picking up "Client Review" task text)
-    max_rows = min(40, df.shape[0])
-    max_cols = min(12, df.shape[1])
-
-    label_map = {
-        "date": ["date", "proposal date"],
-        "client": ["client", "client name"],
-        "project": ["project", "project name"],
-        "location": ["location", "site location", "project location"],
-        "state": ["state"],
-        "size_mw": [
-            "size(mw)", "sizemw", "size mw", "project size (mw)", "project size", "mw",
-            "size (mwac)", "size (mw dc)", "size (mwac/mwdc)", "size (mwac/mw dc)"
-        ],
-    }
-
-    def norm_label(s: str) -> str:
-        return re.sub(r"[\s:()/_-]+", "", s.lower())
-
-    variants = {k: {norm_label(v) for v in vals} for k, vals in label_map.items()}
-    found = {}
-
-    for r in range(max_rows):
-        for c in range(max_cols - 1):
-            cell = df.iat[r, c]
-            if not isinstance(cell, str):
-                continue
-            key = norm_label(cell)
-            for field, opts in variants.items():
-                if key in opts:
-                    val = df.iat[r, c + 1] if (c + 1) < df.shape[1] else None
-                    if pd.notna(val) and field not in found:
-                        found[field] = val
-
-    # Merge discovered values
-    for k in ("date", "client", "project", "location", "state"):
-        if k in found and (found[k] is not None and str(found[k]).strip()):
-            info[k] = found[k]
-
-    # Parse size_mw numerically when possible
-    if "size_mw" in found and found["size_mw"] is not None:
-        raw = str(found["size_mw"]).strip()
-        m = re.search(r"([\d.,]+)", raw)
-        if m:
-            try:
-                info["size_mw"] = float(m.group(1).replace(",", ""))
-            except Exception:
-                info["size_mw"] = raw  # fallback to raw
-        else:
-            info["size_mw"] = raw
-
-    return info
-
-
-
-def build_model_rows(path: str):
-    rows = load_proposal_page_rows(path)
-    rows = enrich_with_details(path, rows)
-
-    # Normalize phases
-    for r in rows:
-        r["phase"] = r.get("phase") or _infer_phase(r["task"]) or "30%"
-
-    # Keep rows with either proposal or detail pricing (> 0)
-    filtered = []
-    for r in rows:
-        pp = r.get("proposal_price") or 0
-        dp = r.get("detail_price") or 0
-        if (pp > 0) or (dp > 0):
-            filtered.append(r)
-    rows = filtered
-
-    # Bucket by category/phase (now includes Substation & BESS)
-    buckets = {
-        "Civil": {p: [] for p in PHASES},
-        "Electrical": {p: [] for p in PHASES},
-        "Structural": {p: [] for p in PHASES},
-        "Substation": {p: [] for p in PHASES},
-        "BESS": {p: [] for p in PHASES},
-    }
-    for r in rows:
-        cat, ph = r["category"], r["phase"]
-        if cat in buckets and ph in buckets[cat]:
-            buckets[cat][ph].append(r)
-
-    info = extract_project_info(path)
-    return buckets, info
-
-
-
-    # Bucket by category/phase
-    buckets = {
-        "Civil": {p: [] for p in PHASES},
-        "Electrical": {p: [] for p in PHASES},
-        "Structural": {p: [] for p in PHASES},
-    }
-    for r in rows:
-        cat, ph = r["category"], r["phase"]
-        if cat in buckets and ph in buckets[cat]:
-            buckets[cat][ph].append(r)
-
-    info = extract_project_info(path)
-    return buckets, info
-
-
-def flatten_to_template_rows(buckets, hours_per_day: float, price_source: str, review_pairs: set):
-    """
-    Flatten into the table format ProposalGenerator expects:
-      - Project Initiation (with Civil/Electrical Due Diligence 1d)
-      - Civil, Electrical, Structural, Substation, BESS (30/60/90/IFC)
-      - Prices:
-          * "proposal" → Proposal Page price
-          * "detail"   → Detail price (fallback to Proposal)
-      - Project Closeout (top-level only)
-    """
-    rows_out = []
-    next_id = 1
-
-    # cross-category predecessor trackers
-    e60_first_task_id = None
-    structural_first_task_applied = False
-
-    def add_item(item_id, name, duration, price, is_milestone, indent, enabled, pred_id, lag, pinned, parent_id):
-        rows_out.append({
-            "ID": item_id,
-            "Name": name,
-            "Duration": int(duration or 0),
-            "Price": int(price or 0),
-            "Is Milestone": bool(is_milestone),
-            "Indent Level": int(indent),
-            "Enabled": bool(enabled),
-            "Predecessor ID": pred_id,
-            "Lag": int(lag or 0),
-            "Is Start Pinned": bool(pinned),
-            "Parent ID": parent_id,
-        })
-
-    # ---- Project Initiation ----
-    pi_id = next_id; next_id += 1
-    add_item(pi_id, "Project Initiation", 0, 0, True, 0, True, None, 0, False, None)
-
-    last_pi_child = None
-    civil_dd_id = None
-    electrical_dd_id = None
-    for name, dur in [
-        ("Deposit & Contract Signed", 0),
-        ("Notice to Proceed", 0),
-        ("Civil Start - Civil Due Diligence", 1),
-        ("Electrical Start - Electrical Due Diligence", 1),
-    ]:
-        tid = next_id; next_id += 1
-        add_item(tid, name, dur, 0, False, 1, True, last_pi_child, 0, False, pi_id)
-        last_pi_child = tid
-        if name.startswith("Civil Start"): civil_dd_id = tid
-        if name.startswith("Electrical Start"): electrical_dd_id = tid
-
-    def build_category(cat_key, cat_label):
-        nonlocal next_id, e60_first_task_id, structural_first_task_applied
-        cat_id = next_id; next_id += 1
-        # disabled until we actually add children
-        add_item(cat_id, cat_label, 0, 0, True, 0, False, None, 0, False, None)
-
-        added_any = False
-
-        def _price_of(t):
-            if cat_key == "Structural":
-                if price_source == "detail":
-                    return (t.get("detail_price") or t.get("proposal_price") or 0)
-                return (t.get("proposal_price") or t.get("detail_price") or 0)
-            if price_source == "detail":
-                return (t.get("detail_price") or t.get("proposal_price") or 0)
-            return (t.get("proposal_price") or t.get("detail_price") or 0)
-
-        def _reorder_for_30(tasks, phase):
-            if phase != "30%":
-                return list(tasks)
-            def score(t):
-                return 0 if "plan set" in (t.get("task", "").lower()) else 1
-            return sorted(tasks, key=score)
-
-        prev_phase_last_id = None
-
-        for phase in PHASES:
-            raw = buckets.get(cat_key, {}).get(phase, [])
-            if not raw:
-                continue
-
-            added_any = True
-            tasks = _reorder_for_30(raw, phase)
-
-            # Phase milestone
-            ms_id = next_id; next_id += 1
-            add_item(ms_id, f"{cat_label} — {phase} Design", 0, 0, True, 1, True, None, 0, False, cat_id)
-
-            first_task_id = None
-            last_task_id = None
-
-            for t in tasks:
-                # Duration: use hours if available (now extended to Substation & BESS)
-                if t.get("hours") is not None and cat_key in ("Civil", "Electrical", "Structural", "Substation", "BESS"):
-                    dur_days = math.ceil((t.get("hours") or 0) / float(hours_per_day))
-                else:
-                    dur_days = 0
-
-                tid = next_id; next_id += 1
-                add_item(
-                    tid,
-                    t["task"],
-                    dur_days,
-                    _price_of(t),
-                    False,
-                    2,              # under phase milestone
-                    True,
-                    last_task_id,   # sequential within phase
-                    0,
-                    False,
-                    ms_id
-                )
-                if first_task_id is None:
-                    first_task_id = tid
-                last_task_id = tid
-
-            # Client Review for the selected pairs (unchanged policy)
-            if (cat_key, phase) in review_pairs:
-                cr_id = next_id; next_id += 1
-                add_item(cr_id, "Client Review", 10, 0, False, 2, True, last_task_id, 0, False, ms_id)
-                last_task_id = cr_id
-
-            # Capture first task of Electrical 60% to link Structural’s first task later
-            if cat_key == "Electrical" and phase == "60%" and first_task_id is not None and e60_first_task_id is None:
-                e60_first_task_id = first_task_id
-            elif cat_key == "Electrical" and phase == "90%" and first_task_id is not None and e60_first_task_id is None:
-                e60_first_task_id = first_task_id  # fallback if 60% missing
-
-            # Wire the first task's predecessor
-            if first_task_id is not None:
-                idx = next(i for i in range(len(rows_out)) if rows_out[i]["ID"] == first_task_id)
-                if cat_key == "Structural" and (not structural_first_task_applied) and e60_first_task_id:
-                    rows_out[idx]["Predecessor ID"] = e60_first_task_id
-                    structural_first_task_applied = True
-                elif prev_phase_last_id is not None:
-                    rows_out[idx]["Predecessor ID"] = prev_phase_last_id
-                else:
-                    # For first phase, tie to Due Diligence where applicable; otherwise last Project Initiation child
-                    if phase == "30%":
-                        if cat_key == "Civil" and civil_dd_id:
-                            rows_out[idx]["Predecessor ID"] = civil_dd_id
-                        elif cat_key == "Electrical" and electrical_dd_id:
-                            rows_out[idx]["Predecessor ID"] = electrical_dd_id
-                        else:
-                            rows_out[idx]["Predecessor ID"] = last_pi_child
-                    else:
-                        rows_out[idx]["Predecessor ID"] = last_pi_child
-
-            prev_phase_last_id = last_task_id
-
-        # enable the top-level only if we added content
-        if added_any:
-            for i in range(len(rows_out) - 1, -1, -1):
-                if rows_out[i]["ID"] == cat_id:
-                    rows_out[i]["Enabled"] = True
-                    break
-
-        return cat_id
-
-    # Default review pairs policy remains the same
-    review_pairs = {("Civil", "30%"), ("Civil", "60%"), ("Electrical", "30%"), ("Electrical", "60%")}
-
-    # Order: Civil → Electrical → Structural → Substation → BESS
-    build_category("Civil", "Civil Engineering")
-    build_category("Electrical", "Electrical Engineering")
-    build_category("Structural", "Structural Engineering")
-    build_category("Substation", "Substation Engineering")  # NEW
-    build_category("BESS", "BESS")                          # NEW
-
-    # ---- Project Closeout ----
-    closeout_id = next_id; next_id += 1
-    add_item(closeout_id, "Project Closeout", 0, 0, True, 0, True, None, 0, False, None)
-
-    return rows_out
-
-
-
-    # Default (non-optional) review pairs per spec
-    review_pairs = {("Civil", "30%"), ("Civil", "60%"), ("Electrical", "30%"), ("Electrical", "60%")}
-
-    # Order: Civil → Electrical → Structural
-    build_category("Civil", "Civil Engineering")
-    build_category("Electrical", "Electrical Engineering")
-    build_category("Structural", "Structural Engineering")
-
-    # ---- Project Closeout (top-level only, empty, 0d) ----
-    closeout_id = next_id; next_id += 1
-    add_item(closeout_id, "Project Closeout", 0, 0, True, 0, True, None, 0, False, None)
-
-    return rows_out
-
-
-
-def push_into_generator(gen: ProposalGenerator, project_info, rows_out):
-    """Replace any existing task tree with the new one and refresh the UI."""
-    # Try to clear any existing Treeview if present
-    try:
-        tree = getattr(gen, "tree", None) or getattr(gen, "treeview", None)
-        if tree is not None and hasattr(tree, "get_children"):
-            for iid in tree.get_children(""):
-                tree.delete(iid)
-    except Exception:
-        pass
-
-    # Reset internal containers
-    gen.template_items = []
-    gen.item_id_map = {}
-    gen.task_counter = 0
-
-    # Rebuild from rows
-    id_to_item = {}
-
-    def mk_item(row):
-        return ProposalItem(
-            name=row["Name"],
-            duration=int(row["Duration"] or 0),
-            price=int(row["Price"] or 0),
-            is_milestone=bool(row["Is Milestone"]),
-            indent_level=int(row["Indent Level"] or 0),
-            item_id=int(row["ID"]),
-        )
-
-    for r in rows_out:
-        id_to_item[r["ID"]] = mk_item(r)
-
-    for r in rows_out:
-        item = id_to_item[r["ID"]]
-        pid = r["Parent ID"]
-        if pid:
-            parent = id_to_item.get(pid)
-            if parent:
-                item.parent = parent
-                parent.children.append(item)
-        pred = r["Predecessor ID"]
-        if pred:
-            item.predecessor_id = int(pred)
-            item.predecessor_type = 'FS'
-            item.lag = int(r.get("Lag") or 0)
-        item.enabled.set(bool(r["Enabled"]))
-
-    gen.template_items = [it for it in id_to_item.values() if it.parent is None]
-    gen.item_id_map = {it.id: it for it in id_to_item.values()}
-    gen.task_counter = max(gen.item_id_map.keys()) if gen.item_id_map else 0
-
-    # Project info
-    if project_info.get("project"):
-        gen.project_name.set(str(project_info["project"]))
-    if project_info.get("client"):
-        gen.company_name.set(str(project_info["client"]))
-    date_cell = project_info.get("date")
-    try:
-        if isinstance(date_cell, (pd.Timestamp, datetime)):
-            dt = pd.to_datetime(date_cell).to_pydatetime()
-            gen.project_start_date.set(dt.strftime("%m/%d/%y"))
-        elif isinstance(date_cell, str) and date_cell.strip():
-            dt = pd.to_datetime(date_cell)
-            gen.project_start_date.set(pd.to_datetime(dt).strftime("%m/%d/%y"))
-    except Exception:
-        pass
-
-    # If ProposalGenerator exposes a UI refresh method, call it
-    for meth in ("rebuild_tree", "refresh_tree", "render_tree", "draw_tree"):
-        if hasattr(gen, meth):
-            try:
-                getattr(gen, meth)()
-                break
-            except Exception:
-                pass
-
-
-# ===================
-# Tkinter App
-# ===================
-
-class App(tk.Tk):
-    def __init__(self):
-        super().__init__()
-        self.title("Proposal Builder (Project Schedule Import)")
-        self.geometry("980x720")
-
-        # Backing engine/UI
-        self.pg = ProposalGenerator(self)
-        self.pg.root = self
-
-        # State
-        self.xlsx_path = tk.StringVar(value="")
-        self.price_source = tk.StringVar(value="proposal")  # or "detail"
-
-        # Top controls
-        ctrl = ttk.Frame(self)
-        ctrl.pack(side=tk.TOP, fill=tk.X, padx=8, pady=8)
-
-        ttk.Label(ctrl, text="Excel (Project Schedule):").grid(row=0, column=0, sticky="w")
-        ttk.Entry(ctrl, textvariable=self.xlsx_path, width=60).grid(row=0, column=1, sticky="we", padx=6)
-        ttk.Button(ctrl, text="Upload Project Schedule", command=self.on_upload_xlsx).grid(row=0, column=2, padx=6)
-
-        ttk.Label(ctrl, text="Pricing:").grid(row=1, column=0, sticky="w")
-        ttk.Radiobutton(ctrl, text="Proposal Page", variable=self.price_source, value="proposal").grid(row=1, column=1, sticky="w")
-        ttk.Radiobutton(ctrl, text="Detail (Civil/Elect/Structural)", variable=self.price_source, value="detail").grid(row=1, column=1, sticky="e")
-
-
-
-        ctrl.columnconfigure(1, weight=1)
-
-    # ---- UI handlers ----
-
-    def on_upload_xlsx(self):
-        path = filedialog.askopenfilename(
-            title="Select Project Schedule (Excel)",
-            filetypes=[
-                ("Excel Macro-Enabled", "*.xlsm"),
-                ("Excel Workbook", "*.xlsx"),
-            ],
-        )
-        if path:
-            self.xlsx_path.set(path)
-            # Auto-parse and overhaul immediately after upload
-            self.on_parse_and_populate()
-
-    def on_parse_and_populate(self):
-        xlsx = self.xlsx_path.get().strip()
-        if not xlsx or not os.path.exists(xlsx):
-            messagebox.showerror("Missing file", "Please choose a Project Schedule workbook (.xlsm or .xlsx).")
-            return
-
-        try:
-            buckets, info = build_model_rows(xlsx)
-            review_pairs = {("Civil", "30%"), ("Civil", "60%"), ("Electrical", "30%"), ("Electrical", "60%")}
-
-            rows_out = flatten_to_template_rows(
-                buckets=buckets,
-                hours_per_day=8.0,
-                price_source=self.price_source.get().strip().lower(),
-                review_pairs=review_pairs,
-            )
-
-            push_into_generator(self.pg, info, rows_out)
-            self.pg.calculate_all_dates()
-
-            messagebox.showinfo("Success", "Project Schedule parsed and task tree replaced.")
-        except Exception as e:
-            messagebox.showerror("Error parsing workbook", str(e))
-
-
-if __name__ == "__main__":
-    App().mainloop()
