+++ utils.py
@@ -1,0 +1,14 @@
+import os
+import sys
+
+# --- HELPER FUNCTION FOR PYINSTALLER ---
+# This function helps find bundled files (like fonts and logos) when running as an .exe
+def resource_path(relative_path):
+    """ Get absolute path to resource, works for dev and for PyInstaller """
+    try:
+        # PyInstaller creates a temp folder and stores path in _MEIPASS
+        base_path = sys._MEIPASS
+    except Exception:
+        base_path = os.path.abspath(".")
+
+    return os.path.join(base_path, relative_path)
